+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * New setting: Ignore locked parts. (Default set to false)
   * Parts with their Locked property set to true will not be converted.
 ### Changed
+* Each setting's description box no longer has a fixed height.
 * Tooltip is slightly taller and now has a shadow.
 * Simplified description for 'Check for updates' and 'Delete part when converted' settings.
 * Moved the error message down to not block the Navbar.